Update README.md
This code should run on Google Colab (the free online version). 
I have created this project during my Data Science course @ Naya College, as a part of OOP (Object Oriented Python) section of the course. The passion in making this project was finding a cure for someones' health issues, by creating a network of people that input their network of products recommendations (including their own health issues and the product that helped them). This is only a taste of what i think can be used in different medicine or cosmetic companies. This way the company will use these network of products' data and recommendations, in a way that will help create a new product or even a new cure for a certain health issue. 

The project's name: Cosmetic Correspondance התאמה קוסמטית

In this project the target is to find a compatible product for a certain user, by using the product's benefit and comparing it to the user's skin concern. The benefit and skin concern must meet a complete comparison, the strings should be the same. From this point of comparison, we can achieve the suitable product per user.

At the beggining of the program, a user needs to prompt his details (containing his skin concern) and can also add a product stating what is the product's benefit, meaning - what skin concern will it improve. Once this is done, the user will be asked if he wants to know what product fits their skin concern, once clicking yes then the corresponding product will be printed to the console stating its ingredient.